requirements.txt
pytest == 7.1.3
allure-pytest==2.11.1
selene == 2.0.0b14
selenium~=4.5.0